Refactor

+++ contracts/MultiSigWallet.sol
@@ -2,33 +2,44 @@
 pragma solidity >=0.6.0 <0.7.0;
 
 contract MultiSigWallet {
-
-    event Deposit(address indexed sender, uint amount, uint balance);
-    event SubmitTransaction(address indexed owner, address indexed to, uint indexed txIndex, uint value, bytes data);
-    event ConfirmTransaction(address indexed owner, uint indexed txIndex);
-    event ExecuteTransaction(address indexed owner, uint indexed txIndex);
-    event RevokeConfirmation(address indexed owner, uint indexed txIndex);
+    event Deposit(address indexed sender, uint256 amount, uint256 balance);
+    event SubmitTransaction(
+        address indexed owner,
+        address indexed to,
+        uint256 indexed txIndex,
+        uint256 value,
+        bytes data
+    );
+    event ConfirmTransaction(address indexed owner, uint256 indexed txIndex);
+    event ExecuteTransaction(address indexed owner, uint256 indexed txIndex);
+    event RevokeConfirmation(address indexed owner, uint256 indexed txIndex);
 
     address[] public owners;
     mapping(address => bool) public isOwner;
-    uint public confirmationsRequired;
+    uint256 public confirmationsRequired;
 
     struct Transaction {
         address to;
-        uint value;
+        uint256 value;
         bytes data;
         bool executed;
         mapping(address => bool) isConfirmed;
-        uint confirmations;
+        uint256 confirmations;
     }
 
     Transaction[] public transactions;
 
-    constructor(address[] memory _owners, uint _confirmationsRequired) public {
+    constructor(address[] memory _owners, uint256 _confirmationsRequired)
+        public
+    {
         require(_owners.length > 0, "Owners required");
-        require(_confirmationsRequired > 0 && _confirmationsRequired <= _owners.length, "Invalid number of required confirmations");
+        require(
+            _confirmationsRequired > 0 &&
+                _confirmationsRequired <= _owners.length,
+            "Invalid number of required confirmations"
+        );
 
-        for (uint i = 0; i < _owners.length; i++) {
+        for (uint256 i = 0; i < _owners.length; i++) {
             address owner = _owners[i];
 
             require(!isOwner[owner], "Duplicate owner");
@@ -40,7 +51,7 @@ contract MultiSigWallet {
         confirmationsRequired = _confirmationsRequired;
     }
 
-    receive () external payable {
+    receive() external payable {
         emit Deposit(msg.sender, msg.value, address(this).balance);
     }
 
@@ -49,35 +60,50 @@ contract MultiSigWallet {
         _;
     }
 
-    function submitTransaction(address _to, uint _value, bytes memory _data) public onlyOwner {
-        uint txIndex = transactions.length;
-        transactions.push(Transaction({
-            to: _to,
-            value: _value,
-            data: _data,
-            executed: false,
-            confirmations: 0
-        }));
+    function submitTransaction(
+        address _to,
+        uint256 _value,
+        bytes memory _data
+    ) public onlyOwner {
+        uint256 txIndex = transactions.length;
+        transactions.push(
+            Transaction({
+                to: _to,
+                value: _value,
+                data: _data,
+                executed: false,
+                confirmations: 0
+            })
+        );
 
         emit SubmitTransaction(msg.sender, _to, txIndex, _value, _data);
     }
 
-    modifier txExists(uint _txIndex) {
+    modifier txExists(uint256 _txIndex) {
         require(_txIndex < transactions.length, "Tx doesn't exist");
         _;
     }
 
-    modifier notExecuted(uint _txIndex) {
+    modifier notExecuted(uint256 _txIndex) {
         require(!transactions[_txIndex].executed, "Tx already executed");
         _;
     }
 
-    modifier notConfirmed(uint _txIndex) {
-        require(!transactions[_txIndex].isConfirmed[msg.sender], "Tx already confirmed");
+    modifier notConfirmed(uint256 _txIndex) {
+        require(
+            !transactions[_txIndex].isConfirmed[msg.sender],
+            "Tx already confirmed"
+        );
         _;
     }
 
-    function confirmTransaction(uint _txIndex) public onlyOwner txExists(_txIndex) notExecuted(_txIndex) notConfirmed(_txIndex) {
+    function confirmTransaction(uint256 _txIndex)
+        public
+        onlyOwner
+        txExists(_txIndex)
+        notExecuted(_txIndex)
+        notConfirmed(_txIndex)
+    {
         Transaction storage transaction = transactions[_txIndex];
         transaction.isConfirmed[msg.sender] = true;
         transaction.confirmations += 1;
@@ -85,18 +111,33 @@ contract MultiSigWallet {
         emit ConfirmTransaction(msg.sender, _txIndex);
     }
 
-    function executeTransaction(uint _txIndex) public onlyOwner txExists(_txIndex) notExecuted(_txIndex) {
+    function executeTransaction(uint256 _txIndex)
+        public
+        onlyOwner
+        txExists(_txIndex)
+        notExecuted(_txIndex)
+    {
         Transaction storage transaction = transactions[_txIndex];
-        require(transaction.confirmations >= confirmationsRequired, "Not enough confirmations");
+        require(
+            transaction.confirmations >= confirmationsRequired,
+            "Not enough confirmations"
+        );
         transaction.executed = true;
 
-        (bool success, ) = transaction.to.call{value: transaction.value}(transaction.data);
+        (bool success, ) = transaction.to.call{value: transaction.value}(
+            transaction.data
+        );
         require(success, "Tx failed");
 
         emit ExecuteTransaction(msg.sender, _txIndex);
     }
 
-    function revokeConfirmation(uint _txIndex) public onlyOwner txExists(_txIndex) notExecuted(_txIndex) {
+    function revokeConfirmation(uint256 _txIndex)
+        public
+        onlyOwner
+        txExists(_txIndex)
+        notExecuted(_txIndex)
+    {
         Transaction storage transaction = transactions[_txIndex];
         require(transaction.isConfirmed[msg.sender], "Tx not confirmed");
 
@@ -106,22 +147,41 @@ contract MultiSigWallet {
         emit RevokeConfirmation(msg.sender, _txIndex);
     }
 
-    function getOwners() public view returns(address[] memory) {
+    function getOwners() public view returns (address[] memory) {
         return owners;
     }
 
-    function getTransactionCount() public view returns(uint) {
+    function getTransactionCount() public view returns (uint256) {
         return transactions.length;
     }
 
-    function getTransaction(uint _txIndex) public view 
-        returns(address to, uint value, bytes memory data, bool executed, uint confirmations) {
-            Transaction storage transaction = transactions[_txIndex];
+    function getTransaction(uint256 _txIndex)
+        public
+        view
+        returns (
+            address to,
+            uint256 value,
+            bytes memory data,
+            bool executed,
+            uint256 confirmations
+        )
+    {
+        Transaction storage transaction = transactions[_txIndex];
 
-            return(transaction.to, transaction.value, transaction.data, transaction.executed, transaction.confirmations);
+        return (
+            transaction.to,
+            transaction.value,
+            transaction.data,
+            transaction.executed,
+            transaction.confirmations
+        );
     }
 
-    function isConfirmed(uint _txIndex, address _owner) public view returns(bool) {
+    function isConfirmed(uint256 _txIndex, address _owner)
+        public
+        view
+        returns (bool)
+    {
         Transaction storage transaction = transactions[_txIndex];
         return transaction.isConfirmed[_owner];
     }
